Preserve the background when converting to png
I can well imagine that many SVGs are designed to be transparent, but I’m not
sure how that’s supposed to work for app icons. In my testing, when my white
background was stripped, the icon was placed on a black background (by iOS)
which didn’t work for my colors. I think it’s probably safer to preserve the
background.